// types/auth.d.ts
import NextAuth, { type DefaultSession } from 'next-auth'

export type AuthActionResponse = {
  success: boolean
  error?: string
  message?: string
}

declare module 'next-auth' {
  /**
   * Returned by `auth`, `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
   */
  interface Session {
    user: {
      /** The user's postal address. */
      role: string
      emailVerified: Date | null
      provider?: 'google' | 'github' | null
      /**
       * By default, TypeScript merges new interface properties and overwrites existing ones.
       * In this case, the default session user properties will be overwritten,
       * with the new ones defined above. To keep the default session user properties,
       * you need to add them back into the newly declared interface.
       */
    } & DefaultSession['user']
  }
}